requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
Werkzeug==2.3.7
SQLAlchemy==2.0.21
python-dotenv==1.0.0
pdfplumber==0.10.2
pytesseract==0.3.10
Pillow==10.0.1
opencv-python==4.8.0.76
cryptography==41.0.3
gunicorn==21.2.0
email-validator==2.0.0
python-dateutil==2.8.2
pytest==7.4.2
